# docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Notification Service",
        "description": "Accept a notification with POST /notifications (HTTP 202). Delivery is asynchronous; poll GET /notifications/{id} for per-channel status and attempt history. GET /users/{userId}/notifications lists that user's notifications, optionally since a date-time.\n",
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Relative to the host serving this application"
        }
    ],
    "paths": {
        "/health": {
            "get": {
                "tags": [
                    "Operations"
                ],
                "summary": "Liveness check",
                "operationId": "get_health",
                "responses": {
                    "200": {
                        "description": "Service is up",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "status"
                                    ],
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "ok"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/notifications": {
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Accept a notification",
                "operationId": "post_notifications_create",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SendNotificationRequest"
                            },
                            "example": {
                                "userId": "user-1",
                                "idempotencyKey": "order-42",
                                "channels": [
                                    "email"
                                ],
                                "subject": "Hello",
                                "body": "Your order is confirmed.",
                                "requiresUserAction": false
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Notification accepted; deliveries are pending",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Notification"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Idempotent replay of an existing notification",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Notification"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Malformed JSON",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Problem"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed, unknown user, or missing contact for a requested channel",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Problem"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/notifications/{id}": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Notification status",
                "operationId": "get_notifications_get",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Current notification and per-channel delivery status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Notification"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Unknown notification id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Problem"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/{userId}/notifications": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Notifications for one user",
                "operationId": "get_user_notifications_list",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "user-1"
                        }
                    },
                    {
                        "name": "since",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2026-01-01T00:00:00+00:00"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notifications created at or after since, newest first",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserNotificationListSchema"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Query parameter since is not a date-time",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Problem"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ProblemViolationSchema": {
                "required": [
                    "propertyPath",
                    "message"
                ],
                "properties": {
                    "propertyPath": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Problem": {
                "required": [
                    "type",
                    "title",
                    "status"
                ],
                "properties": {
                    "type": {
                        "type": "string",
                        "format": "uri",
                        "example": "about:blank"
                    },
                    "title": {
                        "type": "string",
                        "example": "Unknown user"
                    },
                    "status": {
                        "type": "integer",
                        "example": 422
                    },
                    "detail": {
                        "type": "string",
                        "example": "No contact for channel sms",
                        "nullable": true,
                        "default": null
                    },
                    "violations": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProblemViolationSchema"
                        },
                        "default": []
                    }
                },
                "type": "object"
            },
            "AttemptSchema": {
                "required": [
                    "id",
                    "provider",
                    "outcome",
                    "startedAt"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "provider": {
                        "type": "string",
                        "example": "smtp"
                    },
                    "outcome": {
                        "type": "string",
                        "example": "succeeded"
                    },
                    "startedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "finishedAt": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "errorCode": {
                        "type": "string",
                        "example": "invalid_address",
                        "nullable": true
                    },
                    "errorMessage": {
                        "type": "string",
                        "nullable": true
                    },
                    "providerMessageId": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "DeliverySchema": {
                "required": [
                    "id",
                    "channel",
                    "status",
                    "recipient",
                    "attempts"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "channel": {
                        "type": "string",
                        "enum": [
                            "email",
                            "sms"
                        ]
                    },
                    "status": {
                        "type": "string",
                        "example": "pending"
                    },
                    "recipient": {
                        "type": "string",
                        "example": "user1@example.test"
                    },
                    "provider": {
                        "type": "string",
                        "example": "smtp",
                        "nullable": true
                    },
                    "sentAt": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "attempts": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AttemptSchema"
                        }
                    }
                },
                "type": "object"
            },
            "Notification": {
                "required": [
                    "id",
                    "userId",
                    "status",
                    "createdAt",
                    "subject",
                    "body",
                    "deliveries"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "userId": {
                        "type": "string",
                        "example": "user-1"
                    },
                    "status": {
                        "type": "string",
                        "example": "pending"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "subject": {
                        "type": "string",
                        "example": "Hello"
                    },
                    "body": {
                        "type": "string",
                        "example": "Your order is confirmed."
                    },
                    "deliveries": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DeliverySchema"
                        }
                    }
                },
                "type": "object"
            },
            "RecipientOverrideRequest": {
                "properties": {
                    "email": {
                        "type": "string",
                        "nullable": true,
                        "default": null
                    },
                    "phone": {
                        "type": "string",
                        "nullable": true,
                        "default": null
                    }
                },
                "type": "object"
            },
            "SendNotificationRequest": {
                "required": [
                    "userId",
                    "idempotencyKey",
                    "channels",
                    "subject",
                    "body"
                ],
                "properties": {
                    "userId": {
                        "type": "string",
                        "example": "user-1",
                        "maxLength": 64
                    },
                    "idempotencyKey": {
                        "type": "string",
                        "example": "order-42",
                        "maxLength": 128
                    },
                    "channels": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "enum": [
                                "email",
                                "sms"
                            ]
                        },
                        "example": [
                            "email"
                        ],
                        "minItems": 1
                    },
                    "subject": {
                        "type": "string",
                        "example": "Hello",
                        "maxLength": 255
                    },
                    "body": {
                        "type": "string",
                        "example": "Your order is confirmed.",
                        "maxLength": 65535
                    },
                    "requiresUserAction": {
                        "type": "boolean",
                        "example": false,
                        "default": false
                    },
                    "recipient": {
                        "nullable": true,
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/RecipientOverrideRequest"
                            }
                        ],
                        "default": null
                    }
                },
                "type": "object"
            },
            "UserNotificationSchema": {
                "required": [
                    "id",
                    "userId",
                    "status",
                    "createdAt",
                    "subject"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "userId": {
                        "type": "string",
                        "example": "user-1"
                    },
                    "status": {
                        "type": "string",
                        "example": "pending"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "subject": {
                        "type": "string",
                        "example": "Hello"
                    }
                },
                "type": "object"
            },
            "UserNotificationListSchema": {
                "required": [
                    "notifications"
                ],
                "properties": {
                    "notifications": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserNotificationSchema"
                        }
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Operations"
        },
        {
            "name": "Notifications"
        }
    ]
}